pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "SLDvec"
version = "0.0.1"
description = "A single line drawing vectorization library"
readme = "README.md"
dependencies = [
    "fastapi>=0.115.0",
    "Jinja2>=3.0.0",
    "numba>=0.55.0",
    "numpy",
    "pillow>=10.0.0",
    "pybind11>=2.13.0",
    "pydantic>=2.10.0",
    "python-multipart>=0.0.20",
    "scikit-image>=0.20.0",
    "svgwrite>=1.4.0",
    "timm>=1.0.0",
    "typer[all]>=0.12.3",
    "typing-extensions>=4.0.0",
    "uvicorn>=0.30.0",
    "networkx>=3.2",
]
requires-python = ">=3.9,<3.13"

[tool.setuptools.packages.find]
where = ["src"]
include = ["SLDvec*"]

[tool.setuptools.package-data]
SLDvec = ["assets/*"]

[project.scripts]
SLDvec = "SLDvec_cli.cli:app"
